+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 12.0.22 / 2026-04-10
 ==================
 
-* [KUSTOM-76] Fix Shipping amount does not match for order between Kustom and Klarna
+* [KUSTOM-76] Fix Shipping amount does not match for order between Kustom and Magento
 * [KUSTOM-58][KUSTOM-77][KUSTOM-82][KUSTOM-83][KUSTOM-84] Fix Orders instantly cancelled by api because of uncaught exceptions
 
 12.0.21 / 2026-02-24